#19: Create UiWidget layer * refactored size related APIs to use Length * added initial popup support * fixed gwt context to work again * imporved table widget API * first Java8 thoughts

+++ mmm-client/mmm-client-ui/mmm-client-ui-core/mmm-client-ui-core-api/src/main/java/net/sf/mmm/client/ui/api/attribute/AttributeReadPosition.java
@@ -17,7 +17,7 @@ public abstract interface AttributeReadPosition {
    * 
    * @return the x position.
    */
-  int getPositionX();
+  double getPositionX();
 
   /**
    * This method gets the pixel position of the top border of this object on the y-axis (vertical).
@@ -26,6 +26,6 @@ public abstract interface AttributeReadPosition {
    * 
    * @return the y position.
    */
-  int getPositionY();
+  double getPositionY();
 
 }
