+++ mmm-client/mmm-client-ui/mmm-client-ui-core/mmm-client-ui-core-api/src/main/java/net/sf/mmm/client/ui/api/attribute/AttributeReadSizeInPixel.java
@@ -2,8 +2,12 @@
  * http://www.apache.org/licenses/LICENSE-2.0 */
 package net.sf.mmm.client.ui.api.attribute;
 
+
 /**
- * This interface gives read access to the size of an object.
+ * This interface gives read access to the size of an object in
+ * {@link net.sf.mmm.client.ui.api.common.SizeUnit#PIXEL}.
+ * 
+ * @see AttributeReadSize
  * 
  * @author Joerg Hohwiller (hohwille at users.sourceforge.net)
  * @since 1.0.0
@@ -11,17 +15,25 @@ package net.sf.mmm.client.ui.api.attribute;
 public abstract interface AttributeReadSizeInPixel {
 
   /**
-   * This method gets the width of this object in pixels.
+   * This method gets the width of this object in pixels.<br/>
+   * <b>NOTE:</b><br/>
+   * The result will typically be an <code>int</code> value. However, to be compatible with
+   * {@link net.sf.mmm.client.ui.api.common.Length#getAmount()} and with some underlying native toolkits we
+   * use <code>double</code> to prevent potential information loss.
    * 
    * @return the width.
    */
-  int getWidthInPixel();
+  double getWidthInPixel();
 
   /**
-   * This method gets the height of this object in pixels.
+   * This method gets the height of this object in pixels.<br/>
+   * <b>NOTE:</b><br/>
+   * The result will typically be an <code>int</code> value. However, to be compatible with
+   * {@link net.sf.mmm.client.ui.api.common.Length#getAmount()} and with some underlying native toolkits we
+   * use <code>double</code> to prevent potential information loss.
    * 
    * @return the height.
    */
-  int getHeightInPixel();
+  double getHeightInPixel();
 
 }
