+++ mmm-client/mmm-client-ui/mmm-client-ui-widget/mmm-client-ui-widget-impl-web-gwt/src/main/java/net/sf/mmm/client/ui/impl/gwt/UiDisplayGwt.java
@@ -25,7 +25,7 @@ public class UiDisplayGwt extends AbstractUiDisplay {
    * {@inheritDoc}
    */
   @Override
-  public int getWidthInPixel() {
+  public double getWidthInPixel() {
 
     return JavaScriptUtil.getInstance().getScreenWidth();
   }
@@ -34,7 +34,7 @@ public class UiDisplayGwt extends AbstractUiDisplay {
    * {@inheritDoc}
    */
   @Override
-  public int getHeightInPixel() {
+  public double getHeightInPixel() {
 
     return JavaScriptUtil.getInstance().getScreenHeight();
   }
